refactor(backend-automation-runner): split modules
